Added tests for intuitive string comparator
Added unit tests for the intuitive string comparator in order to
demonstrate the new unit testing functionality available after #74. New
tests may be added in the same directory as
`IntuitiveStringComparatorTest` and should be in classes named
`<name of class to be tested>Test`.

Unit tests may be run from the command line using `./gradlew test`. To
run tests from Android Studio, you first need to configure the test
artifact (found in the "Build Variants" window) to be "Unit Tests". The
test package should then be visible in the "Project" window. Tests may
then be run by right-clicking on the test package and selecting
"Run 'Tests in 'org.cirdles.chroni''".