v0.9.4 - Reformatted File Picker and New Version Number The version is now v0.9.4
The File Picker now has the option to create a new folder in the current directory, along with the ability to press and hold a file to delete it or copy it. Also fixed issues within the File Picker such as the very last item being covered up when there are more items than the screen size. Closes #136